+++ tests/sql/multiline/insert.sql
@@ -1,0 +1,4 @@
+-- :name insert
+-- :result :insert
+insert into foobar (foo, bar)
+values (:foo, :bar);