+++ tests/sql/multiline/scalar.sql
@@ -1,0 +1,3 @@
+-- :name scalar
+-- :result :scalar
+select 1 from foobar where :thing = 2;