+++ tests/sql/postgres/insert-returning-star.sql
@@ -1,0 +1,2 @@
+-- :name insert_returning_star :one
+insert into test (foo) values (:foo) returning *;