+++ tests/sql/postgres/insert-returning.sql
@@ -1,0 +1,2 @@
+-- :name insert_returning :insert
+insert into test (foo) values (:foo) returning id;